move template-tag options logic to own file
